requirements.txt
flask==1.1.1
flask-sqlalchemy==2.4.0
flask-migrate==2.5.2
flask_login==0.4.1
flask_wtf==0.14.2
pytz
twilio
getmac
xlsxwriter
ujson==4.3.0
email_validator
websockets==9.1
waitress==1.3.0
openpyxl==2.6.2
requests>=2.22.0
gunicorn==19.9.0
fasteners==0.15
flask-talisman==0.7.0
pymodbus==2.2.0
astral==1.10.1
dateutils==0.6.6
itsdangerous==1.1.0
jinja2==2.11.0
opcua==0.98.9
psutil==5.7.0
pykarbon==1.1.8
comtrade==0.0.3
flask-cors==3.0.10
python3-nmap==1.4.9
dataclasses==0.6
asyncua==0.9.14
opencv-python==4.5.3.56
sqlalchemy==1.3.7
WTForms==2.3.3
werkzeug==0.16.1
markupsafe==2.0.1
boto3==1.21.46
opencv-python-headless==4.5.3.56
python-crontab==3.0.0
configparser>=5.2.0
urllib3==1.25.11
pandas==1.4.4
numpy==1.19
PyQt5==5.13.0     
PyQt5-Qt5==5.15.2     
PyQt5-sip==12.13.0    
PyQtWebEngine==5.12       
PyQtWebEngine-Qt5==5.15.2